Reset grafana password via grafana-cli
The `grafana-cli admin` command was introduced in v 4.1 and allows to
change admin user password via cli.

This will allow us to set new password in the manifest and rotate it
during deployment.

Unfortunatelly `grafana-cli` is a little bit silly now and can only read
config from `$WORKING_DIR/conf/defaults.ini` and as we have grfana.db in
`/var/vcap/store/grafana` we need to point it to it somehow.